+++ public/manifest.json
@@ -3,10 +3,10 @@
   "name": "Create React App Sample",
   "icons": [
     {
-      "src": "favicon.ico",
+      "src": "favicon.png",
       "sizes": "64x64 32x32 24x24 16x16",
       "type": "image/x-icon"
-    }
+    } 
   ],
   "start_url": ".",
   "display": "standalone",
