+++ src/App.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import {
   BrowserRouter as Router,
   Switch,
@@ -21,53 +21,63 @@ import Diets from './pages/Diets';
 import Pages from './pages/Pages';
 import Users from './pages/Users';
 import AdminProfil from './pages/AdminProfil';
-import AdminConnection from './pages/AdminConnection';
-
+import AdminAuth from './pages/AdminAuth';
 const { Content } = Layout;
 
 function App () {
+  const [authenticateUser, setAuthenticateUser] = useState(false);
+  const handleAuthenticate = () => {
+    setTimeout(() => {
+      setAuthenticateUser(true);
+    }, 1000);
+  };
   return (
     <div className='App'>
       <Router>
-        <Layout style={{ minHeight: '100vh' }}>
-          <SideBar />
-          <Layout className='site-layout'>
-            <TopBar />
-            <Content style={{ margin: '0 16px' }}>
-              <Switch>
-                <Route exact path='/connexion' component={AdminConnection} />
-                <privateRoute exact path='/' component={Home} />
-                <privateRoute exact path='/articles' component={Articles} />
-                <privateRoute exact path='/recettes' component={Recipes} />
-                <privateRoute
-                  exact
-                  path='/categories-articles'
-                  component={CategoryArticles}
-                />
-                <privateRoute
-                  exact
-                  path='/categories-recettes'
-                  component={CategoryRecipes}
-                />
-                <privateRoute
-                  exact
-                  path='/ingredients'
-                  component={Ingredients}
-                />
-                <privateRoute exact path='/types-plats' component={Dishes} />
-                <privateRoute exact path='/types-repas' component={Meals} />
-                <privateRoute exact path='/regimes' component={Diets} />
-                <privateRoute exact path='/pages' component={Pages} />
-                <privateRoute exact path='/utilisateurs' component={Users} />
-                <privateRoute
-                  exact
-                  path='/mon-profil'
-                  component={AdminProfil}
-                />
-              </Switch>
-            </Content>
+        <Switch>
+          <Layout style={{ minHeight: '100vh' }}>
+            <Route exact path='/connexion'>
+              <AdminAuth onAuthenticate={handleAuthenticate} />
+            </Route>
+            {authenticateUser &&
+              <><SideBar />
+                <Layout className='site-layout'>
+                  <TopBar />
+                  <Content style={{ margin: '0 16px' }}>
+                    <privateRoute exact path='/' component={Home} />
+                    <privateRoute exact path='/articles' component={Articles} />
+                    <privateRoute exact path='/recettes' component={Recipes} />
+                    <privateRoute
+                      exact
+                      path='/categories-articles'
+                      component={CategoryArticles}
+                    />
+                    <privateRoute
+                      exact
+                      path='/categories-recettes'
+                      component={CategoryRecipes}
+                    />
+                    <privateRoute
+                      exact
+                      path='/ingredients'
+                      component={Ingredients}
+                    />
+                    <privateRoute exact path='/types-plats' component={Dishes} />
+                    <privateRoute exact path='/types-repas' component={Meals} />
+                    <privateRoute exact path='/regimes' component={Diets} />
+                    <privateRoute exact path='/pages' component={Pages} />
+                    <privateRoute exact path='/utilisateurs' component={Users} />
+                    <privateRoute
+                      exact
+                      path='/mon-profil'
+                      component={AdminProfil}
+                    />
+
+                  </Content>
+                </Layout>
+              </>}
           </Layout>
-        </Layout>
+        </Switch>
       </Router>
     </div>
   );
