--- src/components/Add.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const Add = () => {
-  return (
-    <>
-    </>
-  );
-};
-
-export default Add;
