--- src/components/Delete.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const Delete = () => {
-  return (
-    <>
-    </>
-  );
-};
-
-export default Delete;
