--- src/components/See.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const See = () => {
-  return (
-    <>
-    </>
-  );
-};
-
-export default See;
