--- src/components/Update.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const Update = () => {
-  return (
-    <>
-    </>
-  );
-};
-
-export default Update;
