+++ src/pages/Ingredients.js
@@ -1,11 +1,9 @@
 import React from 'react';
-import ModalForm from '../components/ModalForm';
-import FormIngredient from '../components/FormIngredient';
 
 const Ingredients = () => {
   return (
     <>
-      <ModalForm title='Ajouter un ingredient' component={FormIngredient} />
+      <h1> Ingredients</h1>
     </>
   );
 };
